feat(auth): Simplify auth: use only oauth
