+++ app/components/TipTap/Header.vue
@@ -48,7 +48,7 @@ const userMenuItems = computed<DropdownMenuItem[][]>(() => [
       icon: 'i-lucide-log-out',
       onSelect: async () => {
         await auth.logout();
-        await navigateTo('/auth/login');
+        await navigateTo('/auth');
       },
     },
   ],
